feat: task 9 -- extract lineage module with direction enum + error propagation
Move mcp/src/lineage.rs to core/src/lineage/mod.rs. Three fixes applied during
the move:

1. Direction: String -> enum (Ancestors | Descendants | Both) with serde derive.
   traverse_lineage now takes Direction; matches replace string comparisons.
2. LineageResult.chain_valid: bool -> Option<bool>. None = not computed,
   Some(true) = verified, Some(false) = invalid. Default on traversal is None
   since full verification runs separately.
3. DB errors propagate via ?: get_parents / get_children now collect rows into
   rusqlite::Result<Vec<_>> (was filter_map(|r| r.ok())); traverse_lineage no
   longer swallows errors with `if let Ok(...)`. detect_cycle / validate_parents
   keep their Result<_, String> semantic contract.

Added test_db_error_propagation: in-memory DB without schema init, asserts
get_parents and traverse_lineage surface Err instead of panicking/swallowing.
Updated the existing 9 tests to use the enum and Option<bool> (direction +
chain_valid assertions folded into test_traverse_ancestors).

Removed mod lineage; from mcp/src/main.rs and deleted mcp/src/lineage.rs --
no production call sites in mcp yet. pub mod lineage; was already registered
in core/src/lib.rs from the task 1 scaffold.

Verification:
- cargo test -p mnemonic-core lineage:: -> 10 passed
- cargo clippy -p mnemonic-core -- -D warnings -> clean
- cargo build -p mnemonic-mcp -> success
