+++ work/mnemonic-core/decisions.md
@@ -149,3 +149,17 @@ Review details — in JSON files via links. QA report — in logs/working/.
 - `cargo build -p mnemonic-mcp` -> success
 - `cargo test -p mnemonic-mcp` -> 3 passed (unchanged)
 - No real Solana RPC URLs or funded keypairs in tests
+
+## Task 9: Extract lineage module + cleanup
+
+**Status:** Done
+**Commit:** (pending)
+**Agent:** main agent
+**Summary:** Moved `lineage.rs` to `core/src/lineage/mod.rs` with three fixes during the move: (1) `Direction` is now an enum (`Ancestors | Descendants | Both`) with serde derive instead of raw `String`; (2) `LineageResult.chain_valid` is `Option<bool>` with `None` meaning "not computed", `Some(true)` / `Some(false)` otherwise; (3) DB errors propagate via `?` — `get_parents` / `get_children` now collect rows into `rusqlite::Result<Vec<_>>` instead of `filter_map(|r| r.ok())`, and `traverse_lineage` no longer swallows errors with `if let Ok(...)`. `detect_cycle` and `validate_parents` keep their `Result<_, String>` semantic-validation contract unchanged. Updated 9 existing tests to use the enum and `Option<bool>`; added `test_db_error_propagation` (in-memory DB without the schema table, asserts `get_parents` + `traverse_lineage` surface `Err` instead of panicking or swallowing). Removed `mod lineage;` from `mcp/src/main.rs` and deleted `mcp/src/lineage.rs` — no production call sites in mcp yet.
+**Deviations:** None — acceptance called for 9 existing + 1 new = 10 tests; folded the `test_traverse_direction_enum` and `test_chain_valid_option` TDD anchors into the existing `test_traverse_ancestors` to keep the count at 10.
+
+**Verification:**
+- `cargo test -p mnemonic-core lineage::` -> 10 passed (9 existing + 1 new)
+- `cargo clippy -p mnemonic-core -- -D warnings` -> clean
+- `cargo build -p mnemonic-mcp` -> success
+- `grep -rn "mod lineage" mcp/src/` -> empty
