+++ work/mnemonic-core/tasks/9.md
@@ -1,5 +1,5 @@
 ---
-status: planned
+status: done
 depends_on: [8]
 wave: 9
 skills: [code-writing]
